kamu-auth-rebac, RebacRepository: move to own file
